# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"message",
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Sex": {
				"type": "string",
				"enum": [
					"female",
					"male"
				],
				"description": "Enums"
			},
			"CreationUserData": {
				"properties": {
					"pets": {
						"items": {
							"properties": {
								"pet_age": {
									"type": "number",
									"format": "double"
								},
								"pet_name": {
									"type": "string"
								},
								"pet_sex": {
									"$ref": "#/components/schemas/Sex"
								},
								"pet_id": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"pet_age",
								"pet_name",
								"pet_sex",
								"pet_id"
							],
							"type": "object"
						},
						"type": "array"
					},
					"city": {
						"type": "string"
					},
					"sex": {
						"$ref": "#/components/schemas/Sex"
					},
					"description": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"interests": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"vk_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"pets",
					"city",
					"sex",
					"description",
					"age",
					"interests",
					"vk_id"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {}
					}
				}
			}
		}
	},
	"info": {
		"title": "myesportapp.service",
		"version": "1.0.0",
		"description": "Backend special for myESportApp",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "1+1=11"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"msg": {
											"type": "string"
										}
									},
									"required": [
										"msg"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/user/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"headers": {
							"IResponse": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreationUserData"
							}
						}
					}
				}
			}
		},
		"/user/{vk_id}": {
			"get": {
				"operationId": "GetUserID",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"headers": {
							"IResponse": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vk_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}